+++ grackle/__init__.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-__version__ = '0.2.1'
-__update_date__ = '2022-12-26_09:25:27'
+__version__ = '0.2.2'
+__update_date__ = '2022-12-26_10:41:05'