+++ CHANGELOG.md
@@ -15,6 +15,15 @@ _Note: 'Unreleased' section below is used for untagged changes that will be issu
 #### Security
 __BEGIN-CHANGELOG__
 
+### [0.2.2] - 2022-12-26
+#### Added
+ - Currency, account filtering to transaction query methods
+ - Option to exclude repayments, currency selection for MvM
+#### Fixed
+ - File update timestamp wasn't updating on refresh
+ - Logic for `is_posted` was invalid, as gnucash uses date `1969-12-31` for an unposted invoice instead of an empty variable
+ - Timeseries plot failed when it ran out of pre-determined colors
+
 ### [0.2.1] - 2022-12-26
 #### Added
  - Version at bottom of base template